requirements2.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
annotated-types=0.6.0=pypi_0
antlr=4.13.1=ha770c72_1
antlr4-python3-runtime=4.13.1=pypi_0
anyio=4.3.0=pypi_0
atk-1.0=2.36.0=ha1a6a79_0
blinker=1.8.1=pypi_0
boost-cpp=1.82.0=hdb19cb5_2
bzip2=1.0.8=h5eee18b_5
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.3.11=h06a4308_0
cairo=1.16.0=hb05425b_5
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=py39h06a4308_0
dbus=1.13.18=hb2f20db_0
distro=1.9.0=pypi_0
exceptiongroup=1.2.1=pypi_0
expat=2.6.2=h6a678d5_0
flask=3.0.3=pypi_0
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=2.001=hcb22688_0
font-ttf-source-code-pro=2.030=hd3eb1b0_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.14.1=h4c34cd2_2
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
freetype=2.12.1=h4a9f257_0
fribidi=1.0.10=h7b6447c_0
gdk-pixbuf=2.42.10=h5eee18b_0
giflib=5.2.1=h5eee18b_3
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gobject-introspection=1.78.1=py39hd56c82f_0
graphite2=1.3.14=h295c915_1
graphviz=2.50.0=h1b29801_1
gtk2=2.24.33=h73c1081_2
gts=0.7.6=hb67d8dd_3
h11=0.14.0=pypi_0
harfbuzz=4.3.0=hf52aaf7_2
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
icu=73.1=h6a678d5_0
idna=3.7=pypi_0
importlib-metadata=7.1.0=pypi_0
itsdangerous=2.2.0=pypi_0
jinja2=3.1.3=py39h06a4308_0
jpeg=9e=h5eee18b_1
krb5=1.20.1=h143b758_1
lark=1.1.9=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libboost=1.82.0=h109eef0_2
libcurl=8.5.0=h251f7ec_0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=13.2.0=hc881cc4_6
libgd=2.3.3=h6a678d5_3
libglib=2.78.4=hdc74915_0
libgomp=13.2.0=hc881cc4_6
libiconv=1.16=h5eee18b_3
libnghttp2=1.57.0=h2d74bed_0
libpng=1.6.39=h5eee18b_0
librsvg=2.54.4=h36cc946_3
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=13.2.0=h95c4c6d_6
libtiff=4.5.1=h6a678d5_0
libtool=2.4.6=h6a678d5_1009
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxml2=2.10.4=hfdd30dd_2
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.5=pypi_0
ncurses=6.4=h6a678d5_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nl2ltl=0.0.6=pypi_0
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
openai=1.12.0=pypi_0
openjdk=11.0.13=h87a67e3_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.13=h7f8727e_0
pango=1.50.7=h05da053_0
pcre2=10.42=hebb0a14_0
pip=23.3.1=py39h06a4308_0
pixman=0.40.0=h7f8727e_1
poppler=22.12.0=h9614445_3
poppler-data=0.4.11=h06a4308_1
pydantic=2.7.1=pypi_0
pydantic-core=2.18.2=pypi_0
pylogics=0.2.1=pypi_0
python=3.9.19=h955ad1f_0
python_abi=3.9=2_cp39
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
setuptools=68.2.2=py39h06a4308_0
sniffio=1.3.1=pypi_0
spot=2.11.6=py39h3d6467e_1
sqlite=3.45.3=h5eee18b_0
tk=8.6.12=h1ccaba5_0
tqdm=4.66.2=pypi_0
typing-extensions=4.11.0=pypi_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.1=pypi_0
werkzeug=3.0.2=pypi_0
wheel=0.41.2=py39h06a4308_0
xz=5.4.6=h5eee18b_0
zipp=3.18.1=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
